admin controller

+++ server/controller/admin_controller.js
@@ -25,44 +25,49 @@ const create = async (req, res) => {
 // Get all the admins accounts
 const getAll = async (req, res) => {
     try {
-      sortFilter = sort(req.query.sort);
-      recSkipper(req.query.page, req.query.pageSize);
-  
-      const admins = await Admin.find().skip(skip).limit(Number(req.query.pageSize));
-  
-      if (admins.length < 1) {
-        return res.status(404).json({ message: 'Admin not found' });
-      }
-  
-      // Add HATEOAS links for each admin
-      const adminsWithLinks = admins.map(admin => ({
-        ...admin._doc,
-        links: [
-          {
-            rel: 'self',
-            href: `api/v1/admins/${admin.phone}`,
-            type: 'GET'
-          },
-          {
-            rel: 'update',
-            href: `api/v1/admins/${admin.phone}`,
-            type: 'PUT'
-          },
-          {
-            rel: 'delete',
-            href: `api/v1/admins/${admin.phone}`,
-            type: 'DELETE'
-          }
-        ]
-      }));
-  
-      res.status(200).json(adminsWithLinks);
+        sortFilter = sort(req.query.sort)
+        const pageSize = Number(req.query.pageSize) || 10;
+        recSkipper(req.query.page, req.query.pageSize);
+        const admins = await Admin.find().skip(skip).limit(pageSize);
+
+        if (admins.length < 1) {
+            return res.status(404).json({ message: 'admin not found' });
+        }
+
+        if (skip < 0 || pageSize < 0) {
+            return res.status(400).json({ message: 'Invalid query parameters' });
+        }
+
+
+        // Add HATEOAS links for each admin
+        const adminsWithLinks = admins.map(admin => ({
+            ...admin._doc,
+            links: [
+                {
+                    rel: 'self',
+                    href: `api/v1/admins/${admin.phone}`,
+                    type: 'GET'
+                },
+                {
+                    rel: 'update',
+                    href: `api/v1/admins/${admin.phone}`,
+                    type: 'PUT'
+                },
+                {
+                    rel: 'delete',
+                    href: `api/v1/admins/${admin.phone}`,
+                    type: 'DELETE'
+                }
+            ]
+        }));
+
+        res.status(200).json(adminsWithLinks);
     } catch (error) {
-      console.error(error);
-      res.status(500).json({ message: 'Internal Server Error' });
+        console.error(error);
+        res.status(500).json({ message: 'Internal Server Error' });
     }
-  };
-  
+};
+
 
 // Get the admin accocunt
 const getOne = async (req, res) => {
@@ -108,7 +113,7 @@ const update = async (req, res) => {
         await admin.save();
 
         // Send the updated admin data as a response
-        res.status(200).json({ message: 'Information updated successfully'});
+        res.status(200).json({ message: 'Information updated successfully' });
     } catch (error) {
         console.error(error);
         res.status(500).json({ message: 'Internal Server Error' });
